Добавлено 4 задание (исправлен тип данных на double)

+++ src/com/ildar/idea/Zadanie4.java
@@ -6,16 +6,16 @@ public class Zadanie4 {
     public static void main(String[] args) {
         System.out.println("Введите значение x");
         Scanner scannerX = new Scanner(System.in);
-        int numberX = scannerX.nextInt();
+        double numberX = scannerX.nextInt();
         System.out.println("Введите значение y");
         Scanner scannerY = new Scanner(System.in);
-        int numberY = scannerY.nextInt();
+        double numberY = scannerY.nextInt();
         System.out.println("Введите значение z");
         Scanner scannerZ = new Scanner(System.in);
-        int numberZ = scannerZ.nextInt();
-        int sarph = (numberX + numberY + numberZ) / 3;
+        double numberZ = scannerZ.nextInt();
+        double sarph = (numberX + numberY + numberZ) / 3;
         System.out.println(sarph);
-        int A = sarph / 2;
+        double A = sarph / 2;
         int B = (int) Math.floor(A);
         if(B > 3) {
             System.out.println("Программа выполнена корректно");
